requirements.txt
django==2.2.6
djangorestframework==3.10.3
drf-yasg==1.17.0
factory-boy==2.12.0
Faker==2.0.3
flake8==3.7.9
coverage==4.5.4